requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
entrypoints==0.4
fastapi==0.94.0
ffmpy==0.3.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
gradio==3.20.1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
kiwisolver==1.4.4
langchain-core==0.2.1
langchain-google-genai==1.0.5
langsmith==0.1.63
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.6
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
orjson==3.10.3
packaging==23.2
pandas==1.5.3
Pillow==9.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.17
pydantic==1.10.6
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.26.0.post1
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.65.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.0
watchdog==4.0.1
websockets==10.4
yarl==1.8.2
